Admins + some tests

+++ src/internal/api/follow_api.py
@@ -1,4 +1,6 @@
-from flask import jsonify, request
+from flask import request
+from src.internal.utils.status_messages import Status
+from src.internal.utils.access_controller import admin_check, does_user_exist, follow_access_check, super_admin_check, user_access_check
 from src.internal.models.follow import FollowedBy, Follows
 from src.internal.models.user import User
 from src.internal import app
@@ -6,18 +8,20 @@ from src.internal import app
 
 @app.route('/api/v1/follow/create', methods=["POST"])
 def create_follow():
-    """
-    This API creates a new following
-    :param json data:
-    :return:
-    """
+    headers = request.headers
     data = request.get_json()
 
+    if does_user_exist(headers["sender_id"]) is None:
+        return Status.not_loged_in()
+
     try:
-        first_user = User.objects.get(username=data["yourUsername"])
-        second_user = User.objects.get(username=data["targetUsername"])
+        first_user = User.objects.get(id=data["user_id"])
+        second_user = User.objects.get(id=data["target_id"])
     except User.DoesNotExist:
-        return jsonify("A user does not exist")
+        return Status.not_found()
+    
+    if user_access_check(headers["sender_id"], first_user.id):
+        return Status.does_not_have_access()
     
     try:
         Follows.objects.get(user_id=first_user, followed_id=second_user)
@@ -27,41 +31,24 @@ def create_follow():
         followBy = FollowedBy(user_id=second_user, follower_id=first_user)
         follow.save()
         followBy.save()
-        return jsonify("created")
-    
-    return jsonify("User is already following that person")
-
-
-#NOT NEEDED, only for testing
-@app.route('/api/v1/follow/get/<id>', methods=["GET"])
-def get_follow(id):
-    follow = Follows.objects(id=id).first()
-    return jsonify(follow)
-
-
-#NOT NEEDED, not supposed to be allowed to update a follow
-@app.route('/api/v1/follow/update', methods=["PUT"])
-def update_follow():
-    return jsonify("Updated follow. NOT IMPLEMENTED")
+        return Status.created()
+    return Status.already_exists()
 
 
 @app.route('/api/v1/follow/delete', methods=["DELETE"])
 def delete_follow():
-    """
-    Deletes a following connection betew two users
-    :param json data:
-    :return:
-    """
-    data = request.get_json()
+    headers = request.headers
 
-    try:
-        follow = Follows.objects.get(user_id=data["user_id"], followed_id=data["target_id"])
-        followBy = FollowedBy.objects.get(user_id=data["target_id"], follower_id=data["user_id"])
-        follow.delete()
-        followBy.delete()
-        return jsonify("Deleted follow")
-    
-    except Follows.DoesNotExist or FollowedBy.DoesNotExist:
-        return jsonify("User is not following that target")
-    except Exception as e:
-        return jsonify("Error deleting follow")
+    if does_user_exist(headers["sender_id"]) is None:
+        return Status.not_loged_in()
+
+    data = request.get_json()
+    if follow_access_check(headers["sender_id"], data["user_id"], data["target_id"]):
+        return Status.does_not_have_access()
+
+    follow = Follows.objects.get(user_id=data["user_id"], followed_id=data["target_id"])
+    followBy = FollowedBy.objects.get(user_id=data["target_id"], follower_id=data["user_id"])
+    follow.delete()
+    followBy.delete()
+    return Status.deleted()
+
